refactor: refactor chatSubmit
