+++ src/common/chat/Submitmessage/SubmitMessage.module.scss
@@ -1,0 +1,60 @@
+.wrapForm {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+}
+
+.textWrap {
+  position: relative;
+  display: flex;
+  width: 100%;
+  flex-grow: 1;
+
+  font-style: normal;
+  font-weight: normal;
+  font-size: 14px;
+  line-height: 20px;
+}
+
+.fileIcon {
+  position: absolute;
+  cursor: pointer;
+  top: 10px;
+  padding: 5px;
+  right: 0;
+  transform: translateX(-100%);
+}
+
+.singInUpTxt {
+  min-width: 21px;
+  min-height: 18px;
+  margin-left: 25px;
+  border: none;
+  background: none;
+  background-image: url('../../img/icons/send-message.svg');
+  background-repeat: no-repeat;
+  background-size: contain;
+  background-position: center;  
+
+  &:hover {
+    opacity: 0.7;
+  }
+}
+
+.input {
+  height: 3rem;
+  border: none;
+  background: #fff5e3;
+  width: 100%;
+  padding: 10px 40px 15px 15px;
+  border-radius: 15px;
+  resize: none;
+  color: black;
+
+  &:focus {
+    box-shadow: 0 0 0 2pt red;
+    border: 1px solid #fff5e3;
+    outline: none;
+    box-shadow: 0 0 3pt 2pt #ffb11b;
+  }
+}
