--- src/common/chat/submit-message/index.js
@@ -1,3 +1,0 @@
-import all from './submit-message.tsx';
-
-export default all;
